+++ src/express/app.js
@@ -1,4 +1,6 @@
 const express = require('express');
+const mongoose = require('mongoose');
+const config = require('../config');
 const app = express();
 const passport = require('passport');
 const session = require('express-session');
@@ -11,6 +13,8 @@ app.use(express.json());
 app.use(session({ secret: 'sessionsecret' }));
 app.use(passport.initialize());
 app.use(passport.session());
+
+mongoose.connect(`${config.mongodbUrl}/${config.mongodbName}`);
 
 app.use('/', router);
